% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{rho}
\usage{
rho(
  x,
  y = NULL,
  return_dist = FALSE,
  m = 10^4,
  ad_hoc_check = TRUE,
  scale_monotone = TRUE
)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector with compatible length to \code{x}.}

\item{return_dist}{return the simulated distribution of the A-R statistic. Default \code{FALSE}.}

\item{m}{the number of times the A-R algorithm is run.}

\item{ad_hoc_check}{check that the estimate of the statistic is not larger than one. Default \code{TRUE}.}

\item{scale_monotone}{return the standardized coefficient. Default is \code{TRUE}.}
}
\value{
Estimate of the expected value or simulated values of the A-R statistic.
}
\description{
Estimator of the A-R statistic.
}
\examples{
set.seed(1)
n = 100
x = rnorm(n)
y = rnorm(n)
rho(x, y)
d = rho(x, y, return_dist = TRUE)
hist(d, probability = TRUE)

y = (x - mean(x))^2 + rnorm(n, sd = 0.1)
rho(x, y)
d = rho(x, y, return_dist = TRUE)
hist(d, probability = TRUE)
}
