% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_rank.R
\name{rho_rank}
\alias{rho_rank}
\title{rho_rank}
\usage{
rho_rank(
  x,
  y = NULL,
  return_dist = FALSE,
  m = 10^4,
  ad_hoc_check = TRUE,
  scale_monotone = TRUE
)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector with compatible length to \code{x}.}

\item{return_dist}{return the simulated distribution of the A-R statistic. The Default value is \code{FALSE}.}

\item{m}{the number of times the A-R algorithm is run.}

\item{ad_hoc_check}{check that the estimate of the statistic is not larger than one. The Default value is \code{TRUE}.}

\item{scale_monotone}{return the standardized coefficient. The Default value is \code{TRUE}.}
}
\value{
Estimate of the expected value or simulated values of the A-R statistic
}
\description{
Estimator of the A-R statistic using ranks. Faster than \code{rho()} when variables are large. The estimator is the same as the one available in the function \code{rho()} when there are no ties in the data.
}
\examples{
set.seed(1)
n = 1000
x = rnorm(n)
y = rnorm(n)
system.time(r <- rho(x, y))
system.time(r_rank <- rho_rank(x, y))
r; r_rank
# When there are ties,
x = sample(x, replace = TRUE)
rho(x, y); rho_rank(x, y)
}
