% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_test_perm.R
\name{rho_test_perm}
\alias{rho_test_perm}
\title{rho_test_perm}
\usage{
rho_test_perm(x, y = NULL, R = 1000, verbose = FALSE, return_dist = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector with compatible length to \code{x}.}

\item{R}{the number of permutations.}

\item{verbose}{print tracing information. The Default value is \code{FALSE}.}

\item{return_dist}{should the simulated distribution of the A-R statistic be returned. The Default value is \code{TRUE}.}
}
\value{
A list object.
\itemize{
   \item{\code{rho_minus_one_mean}}{ estimate of the expected value of the A-R statistic.}
   \item{\code{rho_minus_one_sample}}{ if \code{return_dist = TRUE} returns simulated values of the A-R statistic.}
   \item{\code{p_value}}{ the test p-value.}
 }
}
\description{
Computes the estimate of the A-R statistic and independence inference using a permutation test.
}
\examples{
set.seed(1)
n = 100
x = rnorm(100)
y = rnorm(100)
res = rho_test_perm(x, y)
res$rho_minus_one_mean
res$p_value
hist(res$rho_minus_one_sample, probability = TRUE)
}
